Move code, and edit some __init__.py files (#403)
* Move code, and edit some __init__.py files

* Fix flake

* Simplify imports
