+++ syne_tune/optimizer/schedulers/searchers/botorch/__init__.py
@@ -10,6 +10,6 @@
 # on an "AS IS" BASIS, WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either
 # express or implied. See the License for the specific language governing
 # permissions and limitations under the License.
-from syne_tune.optimizer.schedulers.searchers.hypertune.hypertune_searcher import (  # noqa: F401
-    HyperTuneSearcher,
+from syne_tune.optimizer.schedulers.searchers.botorch.botorch_searcher import (  # noqa: F401
+    BotorchSearcher,
 )
