+++ syne_tune/optimizer/schedulers/searchers/kde/__init__.py
@@ -10,3 +10,9 @@
 # on an "AS IS" BASIS, WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either
 # express or implied. See the License for the specific language governing
 # permissions and limitations under the License.
+from syne_tune.optimizer.schedulers.searchers.kde.kde_searcher import (  # noqa: F401
+    KernelDensityEstimator,
+)
+from syne_tune.optimizer.schedulers.searchers.kde.multi_fidelity_kde_searcher import (  # noqa: F401
+    MultiFidelityKernelDensityEstimator,
+)
